Serializing long tweets
Added functionality to send_tweet function in twitter.py to allow handling of long messages (greater than 260 characters, which is the max 280 minus a padding of 20 characters for metadata) by splitting them up over multiple tweets and adding an tweet number and total tweet thread length to threaded tweets.

Additionally, the first tweet in a thread will be the only tweet with the hashtags and timestamp.

Will need to test this commit in a real world scenario still. Just need to discuss with the maintainer.