Update Dockerfile/docker-compose.yml
I added some more explicit commands for building the docker image out.

Dockerfile:
Removed the copy of the full build directory because Docker actually will just copy all the files recursively but put them directly in the target directory and not copy the actual folder structure. This should have completely messed up the build process but we'll see why it didn't in a second.

Explicitly copied out needed files

Swapped /app to . for the register-number.sh script because that's the current working dir anyway.

docker-compose.yml:
Mapped the signal_scanner_bot from the build dir to /app/signal_scanner_bot. The reason for this is because the build directory was being mounted like a volume and effectively over-writing everything that the Dockerfile was doing. This was the actual reason my builds were being weird and the reason the COPY scripts/register-number.sh . line wasn't really doing anything.