State file to track autoscan
Added code to create a state file to track the liveness of the autoscan feature across application restarts.

run.py:
Add logging to console and listening group to denote the loaded state of the autoscanner

env.py:
Update the _State class to check for the provided state file on load and set the LISTENING property appropriately.

Add update_listening_status method for the _State class to change the property and update the existence of the file as required.

Add get_listening_status_notice method for the _State class to provide an easy way to return the notification message to be sent to the Signal listener group.

messages.py:
Update code to make use of the new methods of the _State class.