Comments and rehomes and twitter.py revamp and one annoying fix
Adding lots of comments for functions and standardizing where stuff is and providing headers and "line breaks" in code so as to be a little bit more readable. Also making use of the _ prepend for private functions vs public functions.

twitter.py:
Added several functions to dramatically simplify the final send_tweet public function. No major changes just abstracting some stuff out of the main function for readability.

Rehomes:
message_timestamp - since this is exclusively for signal this was moved from the filters module to the signal module.

"Fix":
run.py - add False as default value for debug parameter of cli function. This is mostly so that Visual Studio Code will quit nagging me.