Testing urllib3 error catching
I think these except cases should probably handle most of the skippable errors, ie. urllib3 errors that don't need to be panicked on because the stream can just be restarted. Folding into the long term test with some log messages I should be able to search on easily to confirm it's working or not.