Make radio monitor alerting async
It was kind of async before but now it's ASYNC(tm)

Added aiofiles library and aiohttp library to enable the async downloading of files from OpenMHz that will be uploaded to the specified Signal group. I've also renamed everything to be more generic using "radio_monitor" instead of "swat" including the module name.